Move some client side api code
