+++ src/api/index.js
@@ -1,3 +1,3 @@
-import * as eventAPI from './event.api';
+import * as eventAPI from './events/event.api';
 
 export { eventAPI }; // eslint-disable-line import/prefer-default-export
